requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
PyPDF2
python-docx
python-pptx
weaviate-client>=4.0.0
sentence-transformers
httpx
requests
